requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0
anthropic>=0.39.0
requests>=2.31.0
